+++ packages/emails/components/index.ts
@@ -1,2 +1,2 @@
-export { default as AccountConnected } from "./AccountConnected";
-export { default as RegistrationConfirmation } from "./RegistrationConfirmation";
+export { default as AccountConnected } from "./templates/AccountConnected";
+export { default as RegistrationConfirmation } from "./templates/RegistrationConfirmation";
