+++ packages/emails/components/templates/AccountConnected.tsx
@@ -15,10 +15,8 @@ import {
 	Text,
 } from "@react-email/components";
 import type * as React from "react";
-import c, { emailsConfig } from "config";
-import { DefaultFooter, DefaultHeader, RowLinks } from "./shared";
-
-const baseUrl = process.env.NEXT_PUBLIC_BASE_URL!;
+import c, { emailsConfig } from "config"; 
+import { DefaultFooter, DefaultHeader, GetStartedList, baseUrl } from "./shared";
 
 export default function AccountConnected({ firstName }: { firstName: string }) {
 	return (
@@ -70,34 +68,7 @@ export default function AccountConnected({ firstName }: { firstName: string }) {
 								</Text>
 							</Row>
 						</Section>
-						<ul className="mt-0 pt-0">
-							<li className="mb-20">
-								<strong>
-									Go update your account information.{" "}
-								</strong>
-								Things might have changed since you were last
-								here so be sure to{" "}
-								<Link href={`${baseUrl}/settings`}>
-									update your settings.
-								</Link>
-							</li>
-							<li className="mb-20">
-								<strong>Check out our upcoming events! </strong>
-								{`${c.clubName} is always hosting cool events that give you an opportunity to come learn, have snacks, and meet new people so be sure to check out the upcoming `}
-								<Link href={`${baseUrl}/events`}>events.</Link>
-							</li>
-							<li className="mb-20">
-								<strong>
-									Check out how we bring our websites to life.{" "}
-								</strong>
-								All of our source code for the things we build
-								are open source.{" "}
-								<Link href={`${c.sourceCodeLink}`}>
-									Stop by our Github for more.
-								</Link>
-							</li>
-						</ul>
-
+					<GetStartedList />
 						<Section className="text-center">
 							<Button
 								className="bg-brand rounded-lg px-[18px] py-3 text-white"
@@ -105,20 +76,6 @@ export default function AccountConnected({ firstName }: { firstName: string }) {
 							>
 								Go to your dashboard
 							</Button>
-						</Section>
-						<Section className="mt-45 flex w-full items-center justify-center">
-							<Row className="w-full">
-								{emailsConfig.footerLinks.map((link) => (
-									<Column key={link.name} className="p-4">
-										<Link
-											className="font-bold text-black underline"
-											href={link.href}
-										>
-											{link.name}
-										</Link>{" "}
-									</Column>
-								))}
-							</Row>
 						</Section>
 					</Container>
 					<DefaultFooter />
