+++ packages/emails/package.json
@@ -5,7 +5,7 @@
  "main": "./index.tsx",
  "types": "./index.tsx",
   "scripts": {
-    "dev": "email dev --dir ./components --port 3001"
+    "dev": "email dev --dir ./components/templates --port 3001"
   },
   "keywords": [],
   "author": "",
